requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pyjwt>=2.9
python-multipart>=0.0.9
python-dotenv>=1.0.1

langchain>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.1.10
tiktoken>=0.7.0

openai>=1.35.0

azure-search-documents>=11.5.1
azure-identity>=1.17.0
pypdf>=6.0.0
python-docx=1.2.0
